Moved viewmodel to AboutModel.kt
